Update: search controller
