+++ src/routers/rootRouter.js
@@ -1,6 +1,6 @@
 import express from "express";
-import { postJoin, postLogin } from "../controllers/userController";
-import { home, search } from "../controllers/writingController";
+import { postJoin, postLogin, search } from "../controllers/userController";
+import { home } from "../controllers/writingController";
 
 const rootRouter = express.Router();
